requirements.txt
flask
flask_sqlalchemy
flask_migrate
python-dotenv
psycopg2-binary
gunicorn